🔧 chore: 공통, 부모, 아이 관련으로 구조 분리
