--- AI-Service/app/__init__.py
@@ -1,10 +1,0 @@
-# Makes this directory a package
-
-
-
-
-
-
-
-
-
