+++ AI-Service/child/__init__.py
@@ -1,0 +1,1 @@
+# 아이 관련 로직 모듈
