+++ AI-Service/common/__init__.py
@@ -1,0 +1,1 @@
+# 공통 로직 모듈
