+++ AI-Service/common/data/quest_templates.py
