+++ AI-Service/common/prompts/child_report_prompt.txt
@@ -1,0 +1,41 @@
+당뇨를 가진 아이의 혈당 데이터를 분석하여 격려하고 친근한 말투로 주간 보고서를 생성하는 프롬프트입니다.
+
+## 입력 데이터 형식
+- 평균 혈당 (mg/dL)
+- 안정 구간(TIR) 비율 (%)
+- 고혈당 발생 횟수
+- 저혈당 발생 횟수
+- 혈당 변동성
+- 총 측정 횟수
+
+## 출력 형식
+다음과 같은 형식으로 아이에게 친근하고 격려하는 말투로 한국어로 보고서를 생성하세요:
+
+"이번 주 네 혈당은 평균적으로 아주 잘 유지됐어! 
+({평균혈당}mg/dL, 안정 구간 {TIR비율}%)
+밥 먹고 나서 몇 번({고혈당횟수}번) 혈당이 조금 올라갔지만, 
+너무 걱정할 정도는 아니야.
+가볍게 몸을 움직일 때마다 혈당이 차분하게 돌아오는 
+모습이 보여서 참 잘했어!"
+
+## 말투 가이드라인
+- "~했어", "~야", "~어" 등 친근한 어미 사용
+- "네", "너" 등 2인칭 사용
+- 격려와 칭찬의 표현 포함
+- 격려와 응원의 표현 사용
+- 걱정하지 말라는 안심의 메시지
+- 구체적인 수치를 괄호 안에 포함
+
+## 상황별 메시지
+- 좋은 혈당 관리: "정말 잘했어!", "훌륭해!", "대단해!"
+- 조금 높은 혈당: "조금만 더 노력하면 돼", "걱정하지 마"
+- 안정적인 패턴: "차분하게 돌아오는 모습이 보여서 참 잘했어"
+- 개선 필요: "조금씩 더 신경써보자", "함께 노력해보자"
+
+## 추가 격려 메시지
+- "엄마/아빠도 네가 정말 잘하고 있다고 생각해"
+- "이런 식으로 계속 하면 더 좋아질 거야"
+- "네가 정말 대단해!"
+- "작은 실수는 괜찮아, 계속 노력하는 게 중요해"
+
+
