+++ AI-Service/common/prompts/parent_report_prompt.txt
@@ -1,0 +1,38 @@
+당뇨 환자의 혈당 데이터를 분석하여 주간 요약 보고서를 생성하는 프롬프트입니다.
+
+## 입력 데이터 형식
+- 평균 혈당 (mg/dL)
+- 안정 구간(TIR) 비율 (%)
+- 고혈당 발생 횟수
+- 저혈당 발생 횟수
+- 운동 후 혈당 회복 패턴 분석
+
+## 출력 형식
+다음과 같은 형식으로 한국어로 요약 보고서를 생성하세요:
+
+"이번 주 평균 혈당은 {평균혈당}mg/dL, 안정 구간(TIR)은 {TIR비율}%였습니다.
+점심 식사 후 고혈당이 {고혈당횟수}회 발생했으며 저혈당은 {저혈당횟수}회 발생하지 않았습니다.
+규칙적인 저강도 운동 후 혈당이 안정적으로 회복되는 경향이 확인되었습니다."
+
+## 분석 기준
+- 평균 혈당: 7일간의 평균 혈당 수치
+- TIR (Time in Range): 70-180mg/dL 구간에 머무른 시간의 비율
+- 고혈당: 180mg/dL 이상
+- 저혈당: 70mg/dL 미만
+- 운동 효과: 운동 전후 혈당 변화 패턴 분석
+
+## 추가 분석 요소
+- 혈당 변동성 (표준편차)
+- 식사 후 혈당 상승 패턴
+- 야간 혈당 안정성
+- 주말 vs 평일 패턴 차이
+- 스트레스나 질병으로 인한 혈당 변화
+
+## 출력 예시
+"이번 주 평균 혈당은 142mg/dL, 안정 구간(TIR)은 68%였습니다.
+점심 식사 후 고혈당이 3회 발생했으며 저혈당은 0회 발생하지 않았습니다.
+규칙적인 저강도 운동 후 혈당이 안정적으로 회복되는 경향이 확인되었습니다.
+
+추가로 혈당 변동성이 다소 크게 나타났으며, 주말에는 평일 대비 혈당 관리가 어려운 패턴을 보였습니다."
+
+
