+++ AI-Service/main.py
@@ -1,25 +1,13 @@
-from flask import Flask, request, jsonify
+"""리팩토링된 메인 애플리케이션"""
+
+from flask import Flask
 from flask_cors import CORS
-from app.utils.io import load_text
-from app.services.glucose_service import calculate_glucose_metrics, analyze_glucose
-from app.utils.database_utils import (
-    get_member_info, get_glucose_data, get_weekly_glucose_data,
-    save_quests_to_db, get_quests_by_date
-)
-from app.utils.glucose_utils import (
-    format_glucose_data, calculate_weekly_glucose_summary,
-    get_default_date_range
-)
-from app.utils.quest_utils import generate_llm_quests, select_quests_from_pool
-from app.utils.ai_utils import (
-    format_analyze_prompt, extract_json_from_ai_response,
-    get_default_analysis_result, get_default_child_summary
-)
-from app.db import SessionLocal
-from app.models.database_models import Quest
 from config import config
-from datetime import datetime
-from data.quest_templates import RECORD_QUEST_POOL
+
+# 공통, 아이, 부모 라우트 임포트
+from common.routes import combined_quest, get_quests_api
+from child.routes import child_request_api, child_report_api
+from parent.routes import glucose_analyze, parent_approve_api, glucose_summary_api
 
 app = Flask(__name__)
 
@@ -30,332 +18,23 @@ CORS(app,
      allow_headers=config.CORS_HEADERS,
      supports_credentials=True)
 
+# 공통 라우트
+app.route("/quest", methods=["GET"])(combined_quest)
+app.route("/quest/list", methods=["GET"])(get_quests_api)
 
-@app.route("/quest", methods=["GET"])
-def combined_quest():
-    """혈당 퀘스트와 기록 퀘스트를 통합하여 총 4개 퀘스트 반환"""
-    try:
-        member_id = int(request.args.get("member_id", 1))
-        today = request.args.get("date", datetime.now().strftime("%Y-%m-%d"))
-        
-        # 회원 정보 조회
-        member_info = get_member_info(member_id)
-        if not member_info:
-            return jsonify({"error": "회원을 찾을 수 없습니다."}), 404
-        
-        # 해당 날짜에 이미 퀘스트가 있는지 확인
-        existing_quests = get_quests_by_date(member_id, today)
-        if "error" not in existing_quests and existing_quests.get("total_count", 0) > 0:
-            # 기존 퀘스트가 있으면 기존 퀘스트를 반환
-            quest_list = existing_quests.get("quests", [])
-            result = {}
-            for quest in quest_list:
-                result[quest["quest_title"]] = quest["quest_content"]
-            return jsonify({"result": result})
-        
-        # 혈당 데이터 기반 퀘스트 생성
-        glucose_readings = get_glucose_data(member_id, today)
-        blood_sugar_data = format_glucose_data(glucose_readings)
-        glucose_metrics = calculate_glucose_metrics(blood_sugar_data)
-        
-        # LLM을 사용하여 개인화된 퀘스트 생성
-        result = generate_llm_quests(glucose_metrics, member_info)
-        
-        # 퀘스트를 데이터베이스에 저장
-        save_quests_to_db(member_id, result, today)
-        
-        return jsonify({"result": result})
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
+# 아이 관련 라우트
+app.route("/child/request", methods=["POST"])(child_request_api)
+app.route("/child/report", methods=["GET"])(child_report_api)
 
-
-@app.route("/parent/analyze", methods=["GET"])
-def glucose_analyze():
-    """혈당 종합 분석 및 비서 코칭"""
-    try:
-        member_id = int(request.args.get("member_id", 1))
-        today = datetime.now().strftime("%Y-%m-%d")
-        
-        # 회원 정보 조회
-        member_info = get_member_info(member_id)
-        if not member_info:
-            return jsonify({"error": "회원을 찾을 수 없습니다."}), 404
-        
-        # 혈당 데이터 조회 및 지표 계산
-        readings = get_glucose_data(member_id, today)
-        blood_sugar_data = format_glucose_data(readings)
-        glucose_metrics = calculate_glucose_metrics(blood_sugar_data)
-        avg_glucose = glucose_metrics.get('average_glucose', 0)
-        max_glucose = glucose_metrics.get('max_glucose', 0)
-        min_glucose = glucose_metrics.get('min_glucose', 0)
-        spike_count = glucose_metrics.get('spike_count', 0)
-        glucose_values = [r.glucose_mg_dl for r in readings if r.glucose_mg_dl is not None]
-        
-        # AI 분석 시도
-        prompt = load_text("prompts/parent_analyze_prompt.txt")
-        formatted_prompt = format_analyze_prompt(prompt, avg_glucose, max_glucose, min_glucose, spike_count, len(glucose_values))
-        
-        from app.ai import call_openai_api
-        ai_response = call_openai_api(formatted_prompt, "당신은 혈당 관리를 도와주는 친근한 비서입니다. 반드시 JSON 형식으로만 응답해주세요.")
-        
-        analysis_result = extract_json_from_ai_response(ai_response)
-        
-        # AI 분석 실패 시 기본 응답 반환
-        if not analysis_result:
-            analysis_result = get_default_analysis_result(glucose_values, avg_glucose, max_glucose, spike_count)
-        
-        return jsonify(analysis_result)
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/quest/list", methods=["GET"])
-def get_quests_api():
-    """날짜별 퀘스트 조회 API (LLM 분석 포함)"""
-    try:
-        member_id = int(request.args.get("member_id", 1))
-        date_str = request.args.get("date", datetime.now().strftime("%Y-%m-%d"))
-        
-        result = get_quests_by_date(member_id, date_str)
-        
-        if "error" in result:
-            return jsonify(result), 500
-        
-        # LLM을 사용한 퀘스트 완료율 분석
-        completion_rate = result.get("completion_rate", 0)
-        completed_count = result.get("completed_count", 0)
-        total_count = result.get("total_count", 0)
-        
-        # 완료율에 따른 격려 메시지 생성
-        if completion_rate == 100:
-            encouragement = "완벽해! 모든 퀘스트를 완료했어! 정말 대단해!"
-        elif completion_rate >= 75:
-            encouragement = "거의 다 했어! 조금만 더 노력하면 완벽할 거야!"
-        elif completion_rate >= 50:
-            encouragement = "절반 이상 완료했어! 계속 화이팅이야!"
-        elif completion_rate >= 25:
-            encouragement = "시작이 반이야! 조금씩 꾸준히 해보자!"
-        else:
-            encouragement = "오늘부터 시작해보자! 작은 걸음부터 차근차근!"
-        
-        # 결과에 격려 메시지 추가
-        result["encouragement"] = encouragement
-        result["analysis"] = {
-            "completion_rate": completion_rate,
-            "completed_count": completed_count,
-            "total_count": total_count,
-            "remaining_count": total_count - completed_count
-        }
-        
-        return jsonify(result)
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/child/request", methods=["POST"])
-def child_request_api():
-    """아이 퀘스트 완료 요청 API"""
-    try:
-        data = request.get_json()
-        quest_id = data.get("quest_id")
-        member_id = data.get("member_id")
-        
-        if not quest_id or not member_id:
-            return jsonify({"error": "quest_id와 member_id가 필요합니다."}), 400
-        
-        db = SessionLocal()
-        try:
-            quest = db.query(Quest).filter(
-                Quest.id == quest_id,
-                Quest.member_id == member_id
-            ).first()
-            
-            if not quest:
-                return jsonify({"error": "퀘스트를 찾을 수 없습니다."}), 404
-            
-            if quest.is_completed:
-                return jsonify({"error": "이미 완료된 퀘스트입니다."}), 400
-            
-            quest.is_completed = True
-            quest.approval_status = '요청 중'
-            quest.updated_at = datetime.now()
-            
-            db.commit()
-            
-            return jsonify({"message": "퀘스트가 완료되었습니다. 승인 요청이 전송되었습니다.", "quest_id": quest_id})
-            
-        except Exception as e:
-            db.rollback()
-            return jsonify({"error": str(e)}), 500
-        finally:
-            db.close()
-            
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/parent/approve", methods=["POST"])
-def parent_approve_api():
-    """부모 퀘스트 승인 API"""
-    try:
-        data = request.get_json()
-        quest_id = data.get("quest_id")
-        member_id = data.get("member_id")
-        
-        if not quest_id or not member_id:
-            return jsonify({"error": "quest_id와 member_id가 필요합니다."}), 400
-        
-        db = SessionLocal()
-        try:
-            quest = db.query(Quest).filter(
-                Quest.id == quest_id,
-                Quest.member_id == member_id
-            ).first()
-            
-            if not quest:
-                return jsonify({"error": "퀘스트를 찾을 수 없습니다."}), 404
-            
-            if quest.approval_status != '요청 중':
-                return jsonify({"error": "승인 요청 중인 퀘스트가 아닙니다."}), 400
-            
-            quest.approval_status = '달성 완료'
-            quest.updated_at = datetime.now()
-            
-            db.commit()
-            
-            return jsonify({"message": "퀘스트가 승인 완료되었습니다.", "quest_id": quest_id})
-            
-        except Exception as e:
-            db.rollback()
-            return jsonify({"error": str(e)}), 500
-        finally:
-            db.close()
-            
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/parent/report", methods=["GET"])
-def glucose_summary_api():
-    """주간 혈당 분석 요약 API"""
-    try:
-        member_id = int(request.args.get("member_id", 1))
-        start_date = request.args.get("start_date")
-        end_date = request.args.get("end_date")
-        
-        # 날짜가 제공되지 않으면 최근 7일로 설정
-        if not start_date or not end_date:
-            start_date, end_date = get_default_date_range()
-        
-        # 회원 정보 확인
-        member_info = get_member_info(member_id)
-        if not member_info:
-            return jsonify({"error": "회원을 찾을 수 없습니다."}), 404
-        
-        # 주간 혈당 데이터 조회
-        glucose_data = get_weekly_glucose_data(member_id, start_date, end_date)
-        
-        # 혈당 요약 계산
-        summary = calculate_weekly_glucose_summary(glucose_data)
-        
-        # LLM을 사용한 분석
-        formatted_data = format_glucose_data(glucose_data)
-        glucose_metrics = calculate_glucose_metrics(formatted_data)
-        prompt_text = load_text("prompts/parent_report_prompt.txt")
-        analysis_result = analyze_glucose(glucose_metrics, prompt_text, member_info.get('age'))
-        
-        # LLM 결과에서 요약 텍스트 추출 (안전한 처리)
-        summary_text = ''
-        if isinstance(analysis_result, dict):
-            summary_text = analysis_result.get('summary', '')
-        else:
-            print(f"Unexpected analysis_result type: {type(analysis_result)}")
-            print(f"analysis_result content: {analysis_result}")
-        
-        # LLM 결과가 없으면 기본 템플릿 사용
-        if not summary_text:
-            summary_text = f"이번 주 평균 혈당은 {summary['average_glucose']}mg/dL, 안정 구간(TIR)은 {summary['tir_percentage']}%였습니다.\n"
-            summary_text += f"고혈당이 {summary['hyperglycemia_count']}회 발생했으며 저혈당은 {summary['hypoglycemia_count']}회 발생하지 않았습니다.\n"
-            summary_text += f"규칙적인 저강도 운동 후 혈당이 {summary['exercise_recovery_pattern']}."
-        
-        return jsonify({
-            "summary": summary_text,
-            "data": {
-                "period": f"{start_date} ~ {end_date}",
-                "average_glucose": summary['average_glucose'],
-                "tir_percentage": summary['tir_percentage'],
-                "hyperglycemia_count": summary['hyperglycemia_count'],
-                "hypoglycemia_count": summary['hypoglycemia_count'],
-                "glucose_variability": summary['glucose_variability'],
-                "total_readings": len(glucose_data)
-            }
-        })
-        
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/child/report", methods=["GET"])
-def child_report_api():
-    """아이를 대상으로 한 주간 혈당 분석 보고서 API"""
-    try:
-        member_id = int(request.args.get("member_id", 1))
-        start_date = request.args.get("start_date")
-        end_date = request.args.get("end_date")
-        
-        # 날짜가 제공되지 않으면 최근 7일로 설정
-        if not start_date or not end_date:
-            start_date, end_date = get_default_date_range()
-        
-        # 회원 정보 확인
-        member_info = get_member_info(member_id)
-        if not member_info:
-            return jsonify({"error": "회원을 찾을 수 없습니다."}), 404
-        
-        # 주간 혈당 데이터 조회
-        glucose_data = get_weekly_glucose_data(member_id, start_date, end_date)
-        
-        # 혈당 요약 계산
-        summary = calculate_weekly_glucose_summary(glucose_data)
-        
-        # LLM을 사용한 분석
-        formatted_data = format_glucose_data(glucose_data)
-        glucose_metrics = calculate_glucose_metrics(formatted_data)
-        prompt_text = load_text("prompts/child_report_prompt.txt")
-        analysis_result = analyze_glucose(glucose_metrics, prompt_text, member_info.get('age'))
-        
-        # LLM 결과에서 요약 텍스트 추출 (안전한 처리)
-        summary_text = ''
-        if isinstance(analysis_result, dict):
-            summary_text = analysis_result.get('summary', '')
-        else:
-            print(f"Unexpected analysis_result type: {type(analysis_result)}")
-            print(f"analysis_result content: {analysis_result}")
-        
-        # LLM 결과가 없으면 기본 템플릿 사용
-        if not summary_text:
-            summary_text = get_default_child_summary(summary)
-        
-        return jsonify({
-            "summary": summary_text,
-            "data": {
-                "period": f"{start_date} ~ {end_date}",
-                "average_glucose": summary['average_glucose'],
-                "tir_percentage": summary['tir_percentage'],
-                "hyperglycemia_count": summary['hyperglycemia_count'],
-                "hypoglycemia_count": summary['hypoglycemia_count'],
-                "glucose_variability": summary['glucose_variability'],
-                "total_readings": len(glucose_data)
-            }
-        })
-        
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
+# 부모 관련 라우트
+app.route("/parent/analyze", methods=["GET"])(glucose_analyze)
+app.route("/parent/approve", methods=["POST"])(parent_approve_api)
+app.route("/parent/report", methods=["GET"])(glucose_summary_api)
 
 
 if __name__ == "__main__":
     # 데이터베이스 초기화
-    from app.db import init_db
+    from common.db import init_db
     init_db()
     
     # 서버 시작
