+++ AI-Service/parent/__init__.py
@@ -1,0 +1,1 @@
+# 부모 관련 로직 모듈
